+++ Flowery.NET.Gallery.Browser/Program.cs
@@ -1,13 +1,23 @@
+using System.Runtime.Versioning;
 using System.Threading.Tasks;
 using Avalonia;
 using Avalonia.Browser;
 using Flowery.NET.Gallery;
+using Flowery.NET.Gallery.Browser;
+using Flowery.Services;
+
+[assembly: SupportedOSPlatform("browser")]
 
 internal sealed partial class Program
 {
-    private static Task Main(string[] args) => BuildAvaloniaApp()
+    private static Task Main(string[] args)
+    {
+        StateStorageProvider.Configure(new BrowserStateStorage());
+
+        return BuildAvaloniaApp()
             .WithInterFont()
             .StartBrowserAppAsync("out");
+    }
 
     public static AppBuilder BuildAvaloniaApp()
         => AppBuilder.Configure<App>();